flask/requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2==2.9.3
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
